[Ref] Silence some more 64bit warnings.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@10431 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/Ctrl_seq.cpp
@@ -518,7 +518,7 @@ LRESULT COrderList::OnCustomKeyMsg(WPARAM wParam, LPARAM)
 	case kcOrderlistPat7:
 	case kcOrderlistPat8:
 	case kcOrderlistPat9:
-		EnterPatternNum(wParam - kcOrderlistPat0); return wParam;
+		EnterPatternNum(static_cast<UINT>(wParam) - kcOrderlistPat0); return wParam;
 	case kcOrderlistPatMinus:
 		EnterPatternNum(10); return wParam;
 	case kcOrderlistPatPlus:
